+++ product-service/src/main/java/com/vklovan/productservice/entity/Product.java
@@ -1,6 +1,7 @@
 package com.vklovan.productservice.entity;
 
 import lombok.*;
+import org.bson.types.Decimal128;
 import org.springframework.data.annotation.Id;
 
 import java.math.BigDecimal;
@@ -14,6 +15,6 @@ public class Product {
     private String id;
     private String title;
     private String description;
-    private BigDecimal price;
+    private double price;
 
 }
